requirements.txt
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
oauth2client==4.1.3
zulip==0.6.1
transliterate=1.10.2